rename 'image destroy' to 'image remove', add compat options
Similarly to 7c58f028f, where `instance destroy` got renamed to
`instance remove`, this does the same for `image` subparser, so that the
terminology is in sync between both subcommands. In order to keep
backwards compatibility, `instance/image destroy` is now still supported
in CLI, marked as deprecated. This also changes instance and image modules
API, keeping the `destroy()` methods, printing a deprecation warning and
recommending to use `remove()`.

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D936