instance: avoid race condition in _list_domains, add _find_domain
testcloud was crashing when an instance was destroyed by a third-party
after it has been listed but before its state was queried. This patch
ignores such occurence instead of crashing.

There's a new function `_find_domain` which is more useful when stopping
or removing an instance - we don't need to query through all existing
instances, instead just access exactly the one we want.

Differential Revision: https://phab.qa.fedoraproject.org/D1135